renames forms folder to inputs
